Add basic tests for parseVideoId
